refactor: create command factory class
